Separate constraints from problem data
